ENG-19814: Replay last proposal
When the distributed lock is obtained by a member who is initializing
test to see if the state of the system is ambiguous. The state is
ambiguous when not all other members have posted a result to the
latest proposal. When this is detected replay the last proposal and wait
for all hosts to post a response to that proposal and all participant
nodes to be deleted.