ENG-15452: Do not throw underflow if data is consumed
If there is not enough space in the current output buffer to hold all of
the data in the input buffer only throw an exception if the source
buffer has unread data.

This prevents the witnessed buffer underflow error without requiring the
users of SSLBufferDecrypter to change their behavior since the contract
is unchanged.