KO-553, check existence of background leader migration in PrepareStopNode (#15)
It helps to shorten the detection window for two stop node requests run in short interval.
Previously PrepareStopNode checks whether the partition leader layout is **balanced**, but
it might take tens of seconds to move a partition leader. So the second request has a
fairly large window to not detect that. Now the window is reduced to less than a second.